Add Nft Model
